Fix bugs and code smells detected by SonarLint in TsFile encoder (#10324)

+++ iotdb-core/tsfile/src/main/java/org/apache/iotdb/tsfile/encoding/encoder/IntRleEncoder.java
@@ -101,7 +101,7 @@ public class IntRleEncoder extends RleEncoder<Integer> {
 
     int[] tmpBuffer = new int[TSFileConfig.RLE_MIN_REPEATED_NUM];
     for (int i = 0; i < TSFileConfig.RLE_MIN_REPEATED_NUM; i++) {
-      tmpBuffer[i] = (int) bufferedValues[i];
+      tmpBuffer[i] = bufferedValues[i];
     }
     packer.pack8Values(tmpBuffer, 0, bytes);
     // we'll not write bit-packing group to OutputStream immediately
